+++ examples/View/source/main.cpp
@@ -1,6 +1,7 @@
 #include <Small/All.hpp>
 
 #include <MyScene.hpp>
+#include <Bis.hpp>
 #include <iostream>
 
 int main(int argc, char** argv)
@@ -12,7 +13,7 @@ int main(int argc, char** argv)
     app.setTitle("Hello world!")
         .setFPSLimit(60);
     
-    int id = app.add<MyScene>();
+    int id = app.add<Bis>();
     app.setCurrentScene(id);
 
     app.run();