+++ include/Small/Scenes.hpp
@@ -1,0 +1,17 @@
+/**
+ * @file Scenes.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the Scenes folder
+ * @version 0.1
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_system
+#define sgl_small_system
+
+#include <Small/Scenes/Zone.hpp>
+
+#endif